requirements.txt
attrs
aws-cdk.aws-cloudwatch
aws-cdk.aws-ec2
aws-cdk.aws-iam
aws-cdk.aws-secretsmanager
aws-cdk.core
aws-cdk.cx-api
aws-cdk.region-info
boto3
